backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangGraph / LangChain
langgraph==0.0.26
langchain>=0.1.10
langchain-google-genai==0.0.6

# APIs
googlemaps==4.10.0
httpx==0.26.0

# Snowflake
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1

# Caching (optional)
redis==5.0.1

# Vector Search (optional)
faiss-cpu==1.7.4

# Web Scraping
firecrawl-py==0.0.5

# Auth
python-jose[cryptography]==3.3.0

# Utilities
geojson==3.1.0
nest_asyncio==1.6.0
